[Feat] #71 #78 chat gpt 적용 및 scoring api 추가 작업

+++ src/main/java/gdsc/mju/guessme/domain/quiz/QuizService.java
@@ -14,6 +14,7 @@ import gdsc.mju.guessme.domain.quiz.repository.ScoringRepository;
 import gdsc.mju.guessme.domain.user.entity.User;
 import gdsc.mju.guessme.domain.user.repository.UserRepository;
 import gdsc.mju.guessme.global.infra.gcs.GcsService;
+import gdsc.mju.guessme.global.infra.openai.OpenAIService;
 import gdsc.mju.guessme.global.response.BaseException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.mail.javamail.JavaMailSender;
@@ -37,10 +38,10 @@ public class QuizService {
     private final UserRepository userRepository;
     private final PersonRepository personRepository;
     private final GcsService gcsService;
+    private final OpenAIService openAIService;
     private final JavaMailSender javaMailSender;
     private final SpringTemplateEngine templateEngine;
     private final ScoringRepository scoringRepository;
-
 
     public QuizResDto createQuiz(String username, long personId) {
 
@@ -176,39 +177,45 @@ public class QuizService {
     static ExecutorService executorService = Executors.newFixedThreadPool(5);
 
     @Transactional
-    public Boolean scoring(String username, ScoreReqDto dto) throws IOException {
+    public Boolean scoring(
+            String username,
+            ScoreReqDto dto
+    ) throws IOException, BaseException {
 
         User user = userRepository.findByUserId(username)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
+                .orElseThrow(() -> new BaseException(404, "User not found"));
 
-        String infoValue = dto.getInfoValue(); // 클라에서 갖고 있는 정답
-        Long infoId = dto.getInfoId();
+        String infoKey = dto.getInfoKey(); // 문제
+        String infoValue = dto.getInfoValue(); // 정답
 
         String imageUrl = dto.getImage() != null ?
                 gcsService.uploadFile(dto.getImage()) : null;
 
         String textFromImage = detectText(imageUrl); // 이미지 추출 텍스트
 
-        // 한글로만 입력 받는다고 가정, 두 글자 이상 연속으로 중복되는 경우 정답 처리
-        // 이 코드는 gpt로 변경
-        String overlap = findOverlap(infoValue, textFromImage);
-        // 채점 후 삭제하기
-        String fileUUID = imageUrl.split("/")[4];
-        gcsService.deleteFile(fileUUID);
+        // chat gpt를 통해 채점
+        Boolean correct = scoringWithChatGpt(textFromImage, infoKey, infoValue);
+        // 채점 후 이미지 삭제
+        if (imageUrl != null) {
+            String fileUUID = imageUrl.split("/")[4];
+            gcsService.deleteFile(fileUUID);
+        }
 
-        Scoring scoring = scoringRepository.findByInfoId(infoId);
+        // personId로 person 찾기
+        Person person = personRepository.findById(dto.getPersonId())
+                .orElseThrow(() -> new BaseException(404, "Person not found"));
 
-        if (overlap != null) { // 맞았을 경우
+        Scoring scoring = scoringRepository.findByQuestionAndPerson(infoKey, person);
+
+        if (correct) { // 맞았을 경우
             // 테이블에 있는지 조회
-            if (scoringRepository.existsByInfoId(infoId)) {
+            if (scoring != null) {
+                // 있으면 flag = 0으로 삽입
                 scoring.setWrongFlag(0L);
             } else {
-                Person person = personRepository.findById(dto.getPersonId())
-                        .orElseThrow();
-
-                // 없으면 테이블 삽입
+                // 없으면 새로 삽입
                 scoringRepository.save(Scoring.builder()
-                        .infoId(infoId)
+                        .question(infoKey)
                         .wrongFlag(0L)
                         .person(person)
                         .build());
@@ -216,78 +223,86 @@ public class QuizService {
             return Boolean.TRUE;
         } else { // 틀렸을 경우
             // 테이블에 있는지 조회
-            if (scoringRepository.existsByInfoId(infoId)) {
-                // 있으면 flag++
-                Long flag = scoring.getWrongFlag() + 1;
+            if (scoring != null) {
+                // 있으면 flag 증가
+                long flag = scoring.getWrongFlag() + 1;
                 // if flag == 3 -> 행 지우고 메일 보내기
                 if (flag == 3) {
-                    scoringRepository.deleteByInfoId(infoId);
+                    scoringRepository.deleteByQuestionAndPerson(infoKey, person);
                     // 메일 보내기
-                    executorService.submit(new Runnable() {
-
-                        @Override
-                        public void run() {
-                            MimeMessage message = javaMailSender.createMimeMessage();
-                            try {
-//                                MimeMessageHelper messageHelper = new MimeMessageHelper(message, true, "UTF-8");
-
-                                // 1. 메일 제목 설정
-                                message.setSubject("Guess me! : check on your loved one!");
-
-
-                                // 2. 메일 수신자 설정
-                                String receiver = user.getEmail();
-//                                messageHelper.setTo(receiver);
-                                message.addRecipients(MimeMessage.RecipientType.TO, receiver);
-
-                                // 3. 메일 내용 설정
-//                                messageHelper.setText("The quiz score is too low recently, we recommend you to check on your loved one.");
-                                Context context = new Context();
-                                message.setText(templateEngine.process("mail", context), "utf-8", "html");
-//                                messageHelper.addInline("image1", new ClassPathResource("templates/family.jpg"));
-
-
-                                // 4. 메일 전송
-                                javaMailSender.send(message);
-                            } catch (Exception e) {
-                                System.out.println("error : " + e.toString());
-                            }
-                        }
-                    });
+                    sendEmail(user);
                 } else {
                     scoring.setWrongFlag(flag);
                 }
-
-            } else {
-                // 없으면 continue
             }
 
             return Boolean.FALSE;
         }
     }
 
+    /**
+     * 메일 전송
+     * @param user
+     */
+    private void sendEmail(User user) {
+        executorService.submit(() -> {
+            MimeMessage message = javaMailSender.createMimeMessage();
+            try {
+                // 1. 메일 제목 설정
+                message.setSubject("Guess me! : check on your loved one!");
+                // 2. 메일 수신자 설정
+                message.addRecipients(MimeMessage.RecipientType.TO, user.getEmail());
 
-    public String findOverlap(String str1, String str2) {
-        // 더 긴 문자열 str1로 설정
-        if (str1.length() < str2.length()) {
-            String temp = str1;
-            str1 = str2;
-            str2 = temp;
-        }
+                // 3. 메일 내용 설정
+                Context context = new Context();
+                message.setText(templateEngine.process("mail", context), "utf-8", "html");
 
-        // 더 긴 문자열을 기준으로 루프를 돌면서 서브스트링을 만들어 비교
-        for (int i = str2.length(); i > 0; i--) {
-            for (int j = 0; j <= str2.length() - i; j++) {
-                if (str1.contains(str2.substring(j, j + i))) {
-                    String ans = str2.substring(j, j + i);
-                    if (ans.length() >= 2) {
-                        return ans;
-                    }
-                    return null;
-                }
+                javaMailSender.send(message);
+            } catch (Exception e) {
+                throw new RuntimeException("Failed to send email", e);
             }
+        });
+    }
+
+    /**
+     * chat gpt를 통해 채점
+     * @param textFromImage
+     * @param question
+     * @param answer
+     * @return true or false
+     * @throws BaseException
+     */
+    private Boolean scoringWithChatGpt(String textFromImage, String question, String answer) throws BaseException {
+        String prompt = String.format(
+                "You are a scoring machine now. You can only say yes or no. Here is the situation. " +
+                        "I asked an old man a question about guessing %s, and the correct answer is \"%s\"." +
+                        "The format of the answer is free because the elderly with dementia are the submitters." +
+                        "He submitted \"%s\". If it is correct, please say \"yes,\" otherwise say \"no.\"",
+                question, answer, textFromImage
+        );
+
+        System.out.println("prompt = " + prompt);
+        // return true or false
+        return processResponse(this.openAIService.createCompletion(prompt));
+    }
+
+    private Boolean processResponse(String response) {
+        // Check if the response contains "yes"
+        System.out.println("response = " + response);
+
+        // make response lowercase
+        response = response.toLowerCase();
+
+        if (response.contains("yes")) {
+            return true;
         }
 
-        return null; // 겹치는 단어가 없을 경우 null 반환
+        // Check if the response contains "no"
+        if (response.contains("no")) {
+            return false;
+        }
+
+        // If neither "yes" nor "no" is found, return false as the default value
+        return false;
     }
 }
